LayoutTests/platform/iphone-simulator/accessibility/accessibility-hint.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="stylesheet" href="../../../fast/js/resources/js-test-style.css">
<script src="../../../fast/js/resources/js-test-pre.js"></script>
<script>
var successfullyParsed = false;
</script>
</head>
<body id="body">

<a tabindex=0 id="link" href="#here" title="This is a hint">link label</a>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This makes sure that the title attribute is used as accessibilityHint");

    if (window.accessibilityController) {

        document.getElementById("link").focus();
        var link = accessibilityController.focusedElement;
        shouldBe("link.iphoneHint", "'AXHint: This is a hint'");
    }

    successfullyParsed = true;
</script>

<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
